Handle errors from Radar (#7)
When radar has errors upon submission, it returns a 200 status code with
a json string where `isError` is set to true. Along side that it has a
`message` in the json (for all the errors I've seen). This changes Sonar
to propagate this error in the case that the response value is not a
radar ID.