Move to new attachment upload endpoint (#24)
Previously attachments were uploaded as a multipart form request when
creating the radar. Now they are uploaded using a separate PUT request
with the ID of the newly created radar.

One ambiguity here is how to handle attachment upload failure, since at
this point the radar is already created. Currently this is handled as a
failure of the entire radar creation, but that may need to change later.