Clarify that the error is related to your Apple account (#39)
Got 

```
Precondition failed, try logging in on the web to validate your account
```

today when I tried to post a radar. I double checked everything related to my OpenRadar API key when I tried to post a radar today. It didn't occur to me that the error was related to the Apple account and it was driving me nuts 😄 

Hope this slight variation of the error message helps.